reportAnyErrors: Fix undefined responseHeaderMetadata
This error occurs anytime errors are reported by Client.php, which masks the
original error it was going to report.

The `$repsonseHeaderMetadata` variable is not defined within `reportAnyErrors`,
but there is a parameter named `metadata` which is unused.

Replacing the use of `$repsonseHeaderMetadata` with `$metadata` fixes the
undefined variable problem.